requirements.txt
Django==2.2
django-bootstrap4==1.0.1
django-partial-date==1.2.2
djangorestframework==3.10.3
django-filter==2.2.0